Trait tars::allocator::KeyAllocator

pub trait KeyAllocator: Allocator {
    unsafe fn protect_read(ptr: *mut u8, size: usize);
    unsafe fn protect_write(ptr: *mut u8, size: usize);
    unsafe fn protect_none(ptr: *mut u8, size: usize);
}

Trait for keys allocators

Required Methods

unsafe fn protect_read(ptr: *mut u8, size: usize)

Set memory protection of pages allocated at ptr to read-only.

unsafe fn protect_write(ptr: *mut u8, size: usize)

Set memory protection of pages allocated at ptr to write-only. Note that on most architectures write-only cannot be enforced and also allows read access.

unsafe fn protect_none(ptr: *mut u8, size: usize)

Set memory protection of pages allocated at ptr to prevent any access.