Module tars::malloc

Protected memory allocation

Commonly named functions malloc, calloc, realloc and free, share the same logic and mostly the same interfaces than their posix counterparts. They only add an align parameter for data alignment mostly like posix_memalign. This is intended to adhere a bit more to the interface provided by the alloc::heap module.

malloc_key and realloc_key are a bit different. They share the same interfaces than malloc and realloc but provide additional guarantee expected for modifying their memory protections on their allocated regions.

All regions allocated through these functions must call free to deallocate their memory.

realloc and realloc_key currently always work by explictly making a new copy after having allocated a new object.

Mixing different functions together may lead to undetermined behaviors. For instance if an attempt is made to reallocate memory through realloc_key it is expected the memory must have originally been allocated with malloc_key.

Most errors such as bad pointers provided by callers, as well as unexpected internal errors, integrity errors, are treated as irrecoverable. Therefore unless otherwise specified these functions will panic! on error and the heap will be cleaned-up on stack unwinding as each allocator is instantiated and dedicated to a single thread.

Allocations of size zero are handled by returning a pointer to a static page that can't be read nor written, emitting a termination signal on any attempt.

protect_read, protect_write and protect_none functions may be used to change memory protections on regions allocated through malloc_key and realloc_key exclusively.

This malloc implementation is heavily inspired by OpenBSD's malloc.

Functions

calloc
Allocate memory

free
Free memory

malloc
Allocate memory

malloc_key
Allocate memory and allow changes to memory protections

protect_none
Set memory protection to prevent any access

protect_read
Set memory protection to read-only

protect_write
Set memory protection to write-only

realloc
Reallocate memory

realloc_key
Reallocate memory and allow changes to memory protections