Struct tars::ProtBuf

pub struct ProtBuf<T, A = DefaultBufferAllocator> {
    // some fields omitted
}

A protected Buffer

Fixed-length buffer used to handle sensitive data. Try to minimize data copy and zero-out memory on deallocation.

Such a protected buffer must be associated with a protected memory allocator responsible for allocating and deallocating its memory.

A ProtBuf instance always provides a read/write access to its memory elements. See ProtKey for more controlled access.

// Create a new buffer of bytes
let mut buf: ProtBuf<u8, ProtectedBufferAllocator> = ProtBuf::new_zero(42);

// Or more simply, like this with exactly the same result
let mut buf: ProtBuf8 = ProtBuf::new_zero(42);

assert_eq!(buf[21], 0);

// Use a slice to access and manipulate the underlying memory buffer
my_function(&mut buf);

// Create a new random key
let key: ProtKey<u8, ProtectedKeyAllocator> =
     ProtBuf::new_rand_os(42).into_key();

Methods

impl<T: Copy, A: Allocator> ProtBuf<T, A>

fn len(&self) -> usize

Return the length i.e. the number of elements T of this buffer.

fn new(length: usize) -> ProtBuf<T, A>

New allocated buffer with unitilialized memory.

fn new_zero(length: usize) -> ProtBuf<T, A>

New allocated buffer with its memory initialized with bytes of value zero.

fn new_rand<R: Rng>(length: usize, rng: &mut R) -> ProtBuf<T, A>

Allocate a new buffer of size length and fill it with randomly generated bytes. Use rng as random number generator.

fn new_rand_os(length: usize) -> ProtBuf<T, A>

Allocate a new buffer of size length and fill it with randomly generated bytes. Use an instance of OsRng as random number generator.

fn from_slice(values: &[T]) -> ProtBuf<T, A>

New buffer with elements copied from slice values.

unsafe fn from_raw_parts(buf: *const T, length: usize) -> ProtBuf<T, A>

New buffer from unsafe buffer.

fn from_slices(items: &[&[T]]) -> ProtBuf<T, A>

Build a new instance by concatenating slice items together.

fn from_bufs(items: &[&ProtBuf<T, A>]) -> ProtBuf<T, A>

Build a new instance by concatenating ProtBuf items together.

fn as_mut_slice(&mut self) -> &mut [T]

Return a mutable slice into self.

fn cast_to_slice<U>(&self) -> &[U]

Cast self to another compatible type and return a slice on it.

fn cast_to_slice_mut<U>(&mut self) -> &mut [U]

Cast self to another compatible type and return a mut slice on it.

impl<T: Copy, A: KeyAllocator> ProtBuf<T, A>

fn into_key(self) -> ProtKey<T, A>

Transform self into a protected key ProtKey.

impl<T: FromPrimitive + Copy, A: Allocator> ProtBuf<T, A>

fn from_bytes(bytes: &[u8]) -> ProtBuf<T, A>

New buffer from bytes.