Struct tars::ProtKey

pub struct ProtKey<T: Copy, A: KeyAllocator = DefaultKeyAllocator> {
    // some fields omitted
}

A protected key

Transform a ProtBuf instance into a protected key ProtKey and provide tigher access control on its memory.

By default a ProtKey cannot be read nor written to and will only provide separated accesses with limited scopes. Thus, RAII accessor methods must be used to read and write to a ProtKey. Accessing the underlying key is a bit similar to the way of manipulating an object wrapped in RefCell.

// Instantiate a new buffer initialized with random bytes.
// Same as an usual ProtBuf instance but with a different allocator
let buf_rnd = ProtBuf::<u8, ProtectedKeyAllocator>::new_rand_os(32);

// Until here memory buffer is read/write. Turns-it into a key
let key = ProtKey::new(buf_rnd);

// Or more simply, like this with exactly the same result
let key: ProtKey8 = ProtBuf::new_rand_os(32).into_key();

{
    // Request access in read-mode
    let key_read = key.read();
    let byte = key_read[16];
    // ...
}   // Relinquish its read-access

// Alternative way to read its content
key.read_with(|k| encrypt(&k[..], b"abc"));

// Access it in write-mode
let key_write = key.try_write();
if let Some(mut kw) = key_write {
    kw[16] = 42;
}

Methods

impl<T: Copy, A: KeyAllocator> ProtKey<T, A>

fn new(prot_buf: ProtBuf<T, A>) -> ProtKey<T, A>

Take ownership of prot_buf and transform it into a ProtKey. By default prevent any access.

fn from_buf<B: Allocator>(prot_buf: ProtBuf<T, B>) -> ProtKey<T, A>

Consume and copy prot_buf to force using ProtKey's allocator. If prot_buf already uses a KeyAllocator there is no need to make a copy so directly call the default cstor new instead.

fn read(&self) -> ProtKeyRead<T, A>

Return a wrapper to the key in read mode. This method panic! if this key is already accessed in write mode.

fn try_read(&self) -> Option<ProtKeyRead<T, A>>

Return a wrapper to the key in read mode. Return None if the key is already accessed in write mode.

fn read_with<F>(&self, f: F) where F: FnMut(ProtKeyRead<T, A>)

Access the key in read mode and pass a reference to closure f. The key can only be read during this call. This method will panic! if a read access cannot be acquired on this key.

fn write(&self) -> ProtKeyWrite<T, A>

Return a wrapper to the key in write mode. This method panic! if the key is already currently accessed in read or write mode.

fn try_write(&self) -> Option<ProtKeyWrite<T, A>>

Return a wrapper to the key in write mode. Return None if the key is already accessed in read or write mode.

fn write_with<F>(&self, f: F) where F: FnMut(&mut ProtKeyWrite<T, A>)

Access the key in write mode and pass a reference to closure f. The key can only be writtent during this call. This method will panic! if a write access cannot be acquired on this key.