Struct tars::ProtKeyRead

pub struct ProtKeyRead<'a, T: 'a, A: 'a> {
    // some fields omitted
}

An RAII protected key with read access

This instance is the result of a read request on a ProtKey. If no other similar instance on the same ProtKey exists, raw memory access will be revoked when this instance is destructed.

Methods

impl<'a, T: Copy, A: KeyAllocator> ProtKeyRead<'a, T, A>

fn clone_it(&self) -> ProtKeyRead<T, A>

Clone this instance.