Struct tars::ProtKeyWrite

pub struct ProtKeyWrite<'a, T: 'a, A: 'a> {
    // some fields omitted
}

An RAII protected key with write access

This instance is the result of a write request on a ProtKey. Its raw memory may only be written during the lifetime of this object.